requirements.txt
openai
python-dotenv
pytest
streamlit
langgraph
langchain-openai
pydantic
scikit-learn
plotly
numpy
pandas
